피드백 반영 - 좀 더 역할에 맞는 이름으로 클래스,메서드명 변경 - 메서드 역할에 따라 각 클래스로 이동
