+++ src/main/java/racing/RacingCar.java
@@ -2,7 +2,7 @@ package racing;
 
 import racing.domain.Car;
 import racing.domain.RacingCarRequest;
-import racing.service.RacingCarMake;
+import racing.service.RacingCarGame;
 import racing.service.RacingCarRank;
 import racing.service.RacingCarView;
 
@@ -11,19 +11,19 @@ import java.util.List;
 public class RacingCar {
 
     public static void main(String[] args) {
-        doRacing(RacingCarView.getRacingCarRequest());
+        doRacing(RacingCarView.registerCars());
     }
 
     private static void doRacing(RacingCarRequest racingCarRequest) {
         try {
             //자동차 이름 파싱
-            String[] names = RacingCarMake.parseCarNames(racingCarRequest.getNames());
+            String[] names = RacingCarView.parseCarNames(racingCarRequest.getNames());
 
-            //자동차 리스트 생성
-            List<Car> cars = RacingCarMake.getCars(names);
+            //자동차 생성
+            List<Car> cars = RacingCarGame.createCars(names);
 
             //레이싱 시작
-            RacingCarView.startRacing(cars, racingCarRequest.getMoveCnt());
+            RacingCarGame.startRacing(cars, racingCarRequest.getMoveCnt());
 
             //자동차 순위결정
             List<Car> winners = RacingCarRank.rankCars(cars);
