+++ src/test/java/racing/service/MakeTest.java
@@ -11,50 +11,50 @@ public class MakeTest {
 
     @Test
     public void 정규식() {
-        boolean result = RacingCarMake.checkPattern("abc,asdf,erwe");
+        boolean result = RacingCarView.checkPattern("abc,asdf,erwe");
         assertThat(result).isTrue();
-        result = RacingCarMake.checkPattern("a,b,c,d,e");
+        result = RacingCarView.checkPattern("a,b,c,d,e");
         assertThat(result).isTrue();
-        result = RacingCarMake.checkPattern("a,b");
+        result = RacingCarView.checkPattern("a,b");
         assertThat(result).isTrue();
-        result = RacingCarMake.checkPattern(" a,b");
+        result = RacingCarView.checkPattern(" a,b");
         assertThat(result).isTrue();
-        result = RacingCarMake.checkPattern(" a,b ");
+        result = RacingCarView.checkPattern(" a,b ");
         assertThat(result).isTrue();
-        result = RacingCarMake.checkPattern("  a,b  ");
+        result = RacingCarView.checkPattern("  a,b  ");
         assertThat(result).isTrue();
     }
 
     @Test(expected = RuntimeException.class)
     public void 정규식에러() {
         boolean result;
-        result = RacingCarMake.checkPattern(",bb,cc");
+        result = RacingCarView.checkPattern(",bb,cc");
         assertThat(result).isFalse();
-        result = RacingCarMake.checkPattern("aaa,bbb,");
+        result = RacingCarView.checkPattern("aaa,bbb,");
         assertThat(result).isFalse();
     }
 
     @Test
     public void 자동차생성() {
-        String[] names = RacingCarMake.parseCarNames("aaa,bbb,ccc");
-        List<Car> cars = RacingCarMake.getCars(names);
+        String[] names = RacingCarView.parseCarNames("aaa,bbb,ccc");
+        List<Car> cars = RacingCarGame.createCars(names);
         assertThat(cars.size()).isEqualTo(3);
         assertThat(cars.get(0).getName()).isEqualTo("aaa");
 
-        names = RacingCarMake.parseCarNames("aaa,bbb");
-        cars = RacingCarMake.getCars(names);
+        names = RacingCarView.parseCarNames("aaa,bbb");
+        cars = RacingCarGame.createCars(names);
         assertThat(cars.size()).isEqualTo(2);
         assertThat(cars.get(0).getName()).isEqualTo("aaa");
     }
 
     @Test(expected = RuntimeException.class)
     public void 자동차생성에러() {
-        String[] names = RacingCarMake.parseCarNames("aaa,bbb,");
-        List<Car> cars = RacingCarMake.getCars(names);
+        String[] names = RacingCarView.parseCarNames("aaa,bbb,");
+        List<Car> cars = RacingCarGame.createCars(names);
         assertThat(cars.size()).isEqualTo(0);
 
-        names = RacingCarMake.parseCarNames("aaa,bbb,");
-        cars = RacingCarMake.getCars(names);
+        names = RacingCarView.parseCarNames("aaa,bbb,");
+        cars = RacingCarGame.createCars(names);
         assertThat(cars.size()).isEqualTo(0);
     }
 }
